Format code

+++ ethereum/contracts/Interfaces/IWETH.sol
@@ -2,8 +2,10 @@ pragma solidity 0.8.13;
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
-interface IWETH is IERC20{
+interface IWETH is IERC20 {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
+
+    function transfer(address to, uint256 value) external returns (bool);
+
+    function withdraw(uint256) external;
 }
